fix - move supportedBrowsers.ts into src/utils to fix build issue
