+++ packages/components/src/templates/next/components/internal/UnsupportedBrowserBanner/UnsupportedBrowserBanner.tsx
@@ -2,7 +2,7 @@
 
 import { BiInfoCircle } from "react-icons/bi"
 
-import supportedBrowsers from "../../../../../../supportedBrowsers"
+import supportedBrowsers from "../../../../../utils/supportedBrowsers"
 
 interface UnsupportedBrowserBannerProps {
   userAgent?: string | undefined
