+++ packages/components/src/utils/supportedBrowsers.ts
